qcom: Replace memcpy() With memcpy_toio() for the subsys_pil_tz Driver
Commit 16f70273fff749739d4af588a20df77c79e57854 ("arm64: Import latest memcpy()/memmove() implementation")
introduced the memcpy() implementation that contained significant optimizations,
resulting in faster performance for memcpy() operations on ARM64. It also combined
memcpy() and memmove() functions into one assembly file, memcpy.S. memmove.S was removed.

This is good news for wanting faster performance, but it also exposed another
problem. This implementation broke memcpy() usage for copying data to and from
I/O memory as the implementation causes memory alignment to fall apart on arm64.
This should not be a problem given that we have special functions for this usage
that align memory: memcpy_toio() and memcpy_fromio() respectively.

The subsys_pil_tz driver is one such case. Qualcomm removed the memcpy() function
from this driver starting from 5.4 CAF, but as we are on 4.14 CAF, the memcpy is
still here. As this driver is part of the bootup process, memcpy() triggers a kernel
panic as shown by this log:

[    9.123981 02-24 18:48:32.313] [4:    kworker/4:0:   44] subsys-pil-tz 17300000.qcom,lpass: adsp: loading from 0x000000008be00000 to 0x000000008e000000
[    9.124470 02-24 18:48:32.313] [4:    kworker/4:0:   44] Unable to handle kernel paging request at virtual address ffffff80195b6e9c
[    9.124476 02-24 18:48:32.313] [4:    kworker/4:0:   44] Mem abort info:
[    9.124479 02-24 18:48:32.313] [4:    kworker/4:0:   44]   Exception class = DABT (current EL), IL = 32 bits
[    9.124483 02-24 18:48:32.313] [4:    kworker/4:0:   44]   SET = 0, FnV = 0
[    9.124485 02-24 18:48:32.313] [4:    kworker/4:0:   44]   EA = 0, S1PTW = 0
[    9.124488 02-24 18:48:32.313] [4:    kworker/4:0:   44]   FSC = 33
[    9.124491 02-24 18:48:32.313] [4:    kworker/4:0:   44] Data abort info:
[    9.124494 02-24 18:48:32.313] [4:    kworker/4:0:   44]   ISV = 0, ISS = 0x00000061
[    9.124497 02-24 18:48:32.313] [4:    kworker/4:0:   44]   CM = 0, WnR = 1
[    9.124501 02-24 18:48:32.313] [4:    kworker/4:0:   44] swapper pgtable: 4k pages, 39-bit VAs, pgd = 000000005bbb147b
[    9.124504 02-24 18:48:32.313] [4:    kworker/4:0:   44] [ffffff80195b6e9c] *pgd=00000001f848b003, *pud=00000001f848b003, *pmd=00000001f1962003, *pte=00680000fd075703
[    9.124510 02-24 18:48:32.313] [4:    kworker/4:0:   44] Internal error: Oops: 96000061 [#1] PREEMPT SMP
[    9.124514 02-24 18:48:32.313] [4:    kworker/4:0:   44] Modules linked in:
[    9.124518 02-24 18:48:32.313] [4:    kworker/4:0:   44] Process kworker/4:0 (pid: 44, stack limit = 0x00000000572ebf52)
[    9.124523 02-24 18:48:32.313] [4:    kworker/4:0:   44] CPU: 4 PID: 44 Comm: kworker/4:0 Tainted: G S              4.14.336-perf+ #1
[    9.124526 02-24 18:48:32.313] [4:    kworker/4:0:   44] Hardware name: Qualcomm Technologies, Inc. SM8150 V2 PM8150 LG Electronics, Alpha_lao, rev_1.0 (DT)
[    9.124538 02-24 18:48:32.313] [4:    kworker/4:0:   44] Workqueue: events adsp_load_fw
[    9.124542 02-24 18:48:32.313] [4:    kworker/4:0:   44] task: 000000009ce6658e task.stack: 00000000572ebf52
[    9.124548 02-24 18:48:32.313] [4:    kworker/4:0:   44] pc : __memcpy+0x188/0x230
[    9.124554 02-24 18:48:32.313] [4:    kworker/4:0:   44] lr : pil_init_image_trusted+0x1bc/0x540

As we can see, memcpy() in the function pil_init_image_trusted() accesses RAM,
then tries to copy into I/O memory, but ends up copying it to a forbidden location,
triggering a kernel panic. This is because of alignment issues. So change the
call from memcpy() to memcpy_toio()

[Fixes: 16f70273fff749739d4af588a20df77c79e57854 ("arm64: Import latest memcpy()/memmove() implementation")]
Signed-off-by: Kendrenogen <senabuduh4ckerz@gmail.com>